Exctract common code from psi and fir visualizer classes
