+++ compiler/visualizer/tests/org/jetbrains/kotlin/visualizer/psi/AbstractPsiVisualizer.kt
@@ -5,13 +5,10 @@
 
 package org.jetbrains.kotlin.visualizer.psi
 
-import org.jetbrains.kotlin.checkers.KotlinMultiFileTestWithJava
 import org.jetbrains.kotlin.cli.jvm.compiler.KotlinCoreEnvironment
-import org.jetbrains.kotlin.compiler.visualizer.PsiRenderer
+import org.jetbrains.kotlin.compiler.visualizer.PsiVisualizer
 import org.jetbrains.kotlin.resolve.lazy.JvmResolveUtil
-import org.jetbrains.kotlin.test.ConfigurationKind
 import org.jetbrains.kotlin.test.KotlinTestUtils
-import org.jetbrains.kotlin.test.TestJdkKind
 import org.jetbrains.kotlin.visualizer.AbstractVisualizer
 import java.io.File
 
@@ -20,7 +17,7 @@ abstract class AbstractPsiVisualizer : AbstractVisualizer() {
         val ktFiles = environment.getSourceFiles()
         val analysisResult = JvmResolveUtil.analyze(ktFiles, environment)
 
-        val renderer = PsiRenderer(ktFiles.first(), analysisResult)
+        val renderer = PsiVisualizer(ktFiles.first(), analysisResult)
         val psiRenderResult = renderer.render()
 
         val expectedPath = file.absolutePath.replace(replacement.first, replacement.second)
